+++ CHANGELOG.md
@@ -3,5 +3,9 @@
 
 ## v0.2.1 (14 November 2024)
 
+## Changed
+- The 'tools' folder is now emptied before a general installation to avoid confusion with different tools versions.
+
 ### Fixed
 - Virtual Display Driver download/install issue fixed. #23
+- Workaround applied for the "path does contains spaces" issue. The script will ask to move the program to a folder without spaces if the current path contains any spaces. #17